requirements.txt
-i https://pypi.org/simple
asgiref==3.11.0; python_version >= '3.9'
attrs==25.4.0; python_version >= '3.9'
django==6.0.1; python_version >= '3.12'
django-filter==25.2; python_version >= '3.10'
djangorestframework==3.16.1; python_version >= '3.9'
drf-spectacular==0.29.0; python_version >= '3.7'
gunicorn==23.0.0; python_version >= '3.7'
inflection==0.5.1; python_version >= '3.5'
jsonschema==4.26.0; python_version >= '3.10'
jsonschema-specifications==2025.9.1; python_version >= '3.9'
markdown==3.10.1; python_version >= '3.10'
packaging==26.0; python_version >= '3.8'
psycopg2-binary==2.9.11; python_version >= '3.9'
python-dotenv==1.2.1; python_version >= '3.9'
pyyaml==6.0.3; python_version >= '3.8'
referencing==0.37.0; python_version >= '3.10'
rpds-py==0.30.0; python_version >= '3.10'
sqlparse==0.5.5; python_version >= '3.8'
uritemplate==4.2.0; python_version >= '3.9'
